Update 20:
Continued work on Over the Shoulder:
Units properly show through fog of war. Effect is lost when the player turn ends.
The aura effect visualizations are not properly realized because there is no contextualiztaion. May have to refactor event listeners into RTGameState_Effect and modify to bypass this problem.
Added a GameStateContext for effect application. Not sure if it's needed or if it works.

Updated Daybreak Flame to properly affect units that the player (not the squad) can see
Various tab spacing changes to other classes